# src/main/resources/static/openapi/contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "현장뛰자 API — 계약",
    "version": "0.0.1",
...
  },
...
  "paths": {
...
    "/api/explore/cards": {
      "post": {
        "tags": [
          "탐색"
        ],
        "operationId": "getExploreCards",
        "summary": "'하고 싶은 일' 카드(지원할 수 있는 직무의 하는 일, 기관 이름 가림, ADR-0031)",
        "x-auth": "STUDENT",
        "description": "권한: **STUDENT** · 화면: 탐색 · 규칙은 [docs/api/README.md](https://github.com/skuniv-team11/Backend/blob/develop/docs/api/README.md) '엔드포인트별 규칙'",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileBody"
              },
              "examples": {
                "profile-body.request.json": {
                  "summary": "docs/api/profile-body.request.json",
                  "value": {
                    "profile": {
                      "departmentId": 31,
                      "grade": 3,
                      "completedSemesters": 5,
                      "gpa": 3.4,
                      "graduationExpected": false,
                      "interestText": "뷰티 브랜드 SNS 마케팅",
                      "homeAreaCode": "11350",
                      "certificates": []
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploreCards"
                },
                "examples": {
                  "explore-cards.json": {
                    "summary": "docs/api/explore-cards.json",
                    "value": {
                      "round": {
                        "id": 1,
                        "termCode": "2026-2"
                      },
                      "candidates": {
                        "total": 23,
                        "eligible": 21,
                        "needsCheck": 2
                      },
                      "cards": [
                        {
                          "id": "101-4",
                          "text": "(가상) 자사 SNS 채널 콘텐츠 운영 지원"
                        },
                        {
                          "id": "102-2",
                          "text": "(가상) 인플루언서 소통 및 관리"
                        },
                        {
                          "id": "103-1",
                          "text": "(가상) 브랜드 숏폼 콘텐츠 기획"
                        },
                        {
                          "id": "104-3",
                          "text": "(가상) 기획전 제작 및 운영"
                        },
                        {
                          "id": "105-1",
                          "text": "(가상) 판매 데이터 정리와 매출 분석 보조"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "INVALID_INPUT": {
                    "value": {
                      "code": "INVALID_INPUT",
                      "message": "형식·범위 위반. `fields: [{field, reason}]`를 준다",
                      "fields": [
                        {
                          "field": "profile.gpa",
                          "reason": "0.0~4.5, 소수 첫째 자리까지"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED · TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AUTH_REQUIRED": {
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "토큰 없음·잘못된 토큰"
                    }
                  },
                  "TOKEN_EXPIRED": {
                    "value": {
                      "code": "TOKEN_EXPIRED",
                      "message": "토큰 만료(체험 계정은 계정도 지워짐)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN_ROLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "FORBIDDEN_ROLE": {
                    "value": {
                      "code": "FORBIDDEN_ROLE",
                      "message": "역할이 맞지 않음(학생이 현황판 등)"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/explore": {
      "post": {
        "tags": [
          "탐색"
        ],
        "operationId": "explore",
        "summary": "경험 글·카드로 직무 탐색(AI 순서·근거 + 상위 3곳 '왜 맞나요', 실패하면 규칙 추천)",
        "x-auth": "STUDENT",
        "description": "권한: **STUDENT** · 화면: 탐색 · 규칙은 [docs/api/README.md](https://github.com/skuniv-team11/Backend/blob/develop/docs/api/README.md) '엔드포인트별 규칙'",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExploreRequest"
              },
              "examples": {
                "explore.request.json": {
                  "summary": "docs/api/explore.request.json",
                  "value": {
                    "profile": {
                      "departmentId": 31,
                      "grade": 3,
                      "completedSemesters": 5,
                      "gpa": 3.4,
                      "graduationExpected": false,
                      "interestText": "뷰티 브랜드 SNS 마케팅",
                      "homeAreaCode": "11350",
                      "certificates": []
                    },
                    "experiences": [
                      "학과 졸업작품 홍보용 인스타그램 계정을 1년 동안 운영하면서, 게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교했어요.",
                      "화장품 매장 아르바이트에서 손님이 자주 묻는 질문을 메모해 두고 진열을 바꿔 봤어요."
                    ],
                    "cardIds": [
                      "101-4",
                      "102-2",
                      "103-1"
                    ],
                    "consent": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explore"
                },
                "examples": {
                  "explore.json": {
                    "summary": "docs/api/explore.json",
                    "value": {
                      "runId": 12,
                      "createdAt": "2026-10-10T15:20:00+09:00",
                      "round": {
                        "id": 1,
                        "termCode": "2026-2"
                      },
                      "source": "AI",
                      "fallbackReason": null,
                      "judgedWith": "RUN_PROFILE",
                      "hiddenCount": 0,
                      "input": {
                        "experiences": [
                          "학과 졸업작품 홍보용 인스타그램 계정을 1년 동안 운영하면서, 게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교했어요.",
                          "화장품 매장 아르바이트에서 손님이 자주 묻는 질문을 메모해 두고 진열을 바꿔 봤어요."
                        ],
                        "cards": [
                          "(가상) 자사 SNS 채널 콘텐츠 운영 지원",
                          "(가상) 인플루언서 소통 및 관리",
                          "(가상) 브랜드 숏폼 콘텐츠 기획"
                        ],
                        "interestText": "뷰티 브랜드 SNS 마케팅"
                      },
                      "candidates": {
                        "total": 23,
                        "eligible": 21,
                        "needsCheck": 2
                      },
                      "items": [
                        {
                          "rank": 1,
                          "jobId": 101,
                          "title": "국내 마케팅",
                          "team": "마케팅팀",
                          "institution": {
                            "id": 3,
                            "name": "누리뷰티(가상)",
                            "logoPath": "/logos/3.png"
                          },
                          "verdict": "ELIGIBLE",
                          "fit": "STRONG",
                          "evidence": {
                            "studentQuote": "졸업작품 홍보용 인스타그램 계정을 1년 동안 운영하면서",
                            "jobQuote": "(가상) 자사 SNS 채널 콘텐츠 운영 지원",
                            "documentTitle": "누리뷰티(가상) 운영계획서",
                            "page": 3,
                            "reason": "인스타그램 계정을 1년 동안 운영하며 반응을 비교한 경험이 SNS 채널 운영 지원과 바로 이어져요."
                          },
                          "why": {
                            "summary": "인스타그램을 1년 동안 운영하며 반응을 비교해 본 경험이 이 자리의 SNS 채널 운영과 바로 이어져요.",
                            "points": [
                              {
                                "text": "게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교한 경험은, 채널 콘텐츠를 운영하며 반응을 살피는 일에 바로 쓸 수 있어요.",
                                "studentQuote": "게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교했어요",
                                "jobQuote": "(가상) 자사 SNS 채널 콘텐츠 운영 지원"
                              },
                              {
                                "text": "손님이 자주 묻는 질문을 모아 진열을 바꿔 본 경험은, 소비자가 궁금해하는 점에서 상품 아이디어를 내는 일에 도움이 돼요.",
                                "studentQuote": "손님이 자주 묻는 질문을 메모해 두고 진열을 바꿔 봤어요",
                                "jobQuote": "(가상) 트렌드에 맞는 신제품 상품 기획"
                              }
                            ],
                            "tryNew": [
                              {
                                "text": "시장과 경쟁사 제품을 비교해 상품 기획 아이디어를 제안하는 일을 해 볼 수 있어요.",
                                "jobQuote": "(가상) 경쟁사 분석을 통한 상품 기획 아이디어 제안"
                              }
                            ],
                            "prepare": {
                              "text": "운영한 계정에서 저장 수가 가장 높았던 게시물과 무엇을 바꿨는지를 한 장으로 정리해 두면 좋아요.",
                              "jobQuote": "(가상) 자사 SNS 채널 콘텐츠 운영 지원"
                            }
                          }
                        },
                        {
                          "rank": 2,
                          "jobId": 103,
                          "title": "영상콘텐츠 기획",
                          "team": "콘텐츠팀",
                          "institution": {
                            "id": 5,
                            "name": "한빛코스메틱(가상)",
                            "logoPath": "/logos/5.png"
                          },
                          "verdict": "NEEDS_CHECK",
                          "fit": "STRONG",
                          "evidence": {
                            "studentQuote": "게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교했어요",
                            "jobQuote": "(가상) 브랜드 숏폼 콘텐츠 기획",
                            "documentTitle": "한빛코스메틱(가상) 운영계획서",
                            "page": null,
                            "reason": "게시 요소를 바꿔 가며 반응을 비교한 경험이 숏폼 콘텐츠를 기획하고 고치는 일에 쓰여요."
                          },
                          "why": {
                            "summary": "계정을 운영하며 콘텐츠를 만들어 본 경험이 숏폼 콘텐츠 기획으로 이어져요.",
                            "points": [
                              {
                                "text": "게시 요소를 바꿔 가며 저장 수를 비교한 경험은, 숏폼 콘텐츠를 기획하고 반응을 보며 고치는 일에 도움이 돼요.",
                                "studentQuote": "게시 시간과 첫 문장을 바꿔 가며 저장 수를 비교했어요",
                                "jobQuote": "(가상) 브랜드 숏폼 콘텐츠 기획"
                              }
                            ],
                            "tryNew": [
                              {
                                "text": "카메라로 직접 찍고 편집 프로그램으로 영상을 다듬는 일을 새로 해 볼 수 있어요.",
                                "jobQuote": "(가상) 촬영 실습과 영상 편집"
                              }
                            ],
                            "prepare": null
                          }
                        },
                        {
                          "rank": 3,
                          "jobId": 104,
                          "title": "AMD",
                          "team": "세일즈팀",
                          "institution": {
                            "id": 5,
                            "name": "한빛코스메틱(가상)",
                            "logoPath": "/logos/5.png"
                          },
                          "verdict": "ELIGIBLE",
                          "fit": "GOOD",
                          "evidence": {
                            "studentQuote": "화장품 매장 아르바이트에서 손님이 자주 묻는 질문을 메모해 두고",
                            "jobQuote": "(가상) 기획전 제작 및 운영",
                            "documentTitle": "한빛코스메틱(가상) 운영계획서",
                            "page": 4,
                            "reason": "손님 질문을 모아 진열을 바꿔 본 경험이 기획전을 만들고 운영하는 일과 이어져요."
                          },
                          "why": null
                        },
                        {
                          "rank": 4,
                          "jobId": 102,
                          "title": "인플루언서 마케팅",
                          "team": "마케팅팀",
                          "institution": {
                            "id": 3,
                            "name": "누리뷰티(가상)",
                            "logoPath": "/logos/3.png"
                          },
                          "verdict": "ELIGIBLE",
                          "fit": "GOOD",
                          "evidence": {
                            "studentQuote": "학과 졸업작품 홍보용 인스타그램 계정을 1년 동안 운영하면서",
                            "jobQuote": "(가상) 인플루언서 소통 및 관리",
                            "documentTitle": "누리뷰티(가상) 운영계획서",
                            "page": null,
                            "reason": "인스타그램 계정을 운영한 경험이 인플루언서와 소통하며 SNS 채널을 관리하는 일에 쓰여요."
                          },
                          "why": null
                        }
                      ],
                      "blockedBy": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT · CONSENT_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "INVALID_INPUT": {
                    "value": {
                      "code": "INVALID_INPUT",
                      "message": "형식·범위 위반. `fields: [{field, reason}]`를 준다",
                      "fields": [
                        {
                          "field": "profile.gpa",
                          "reason": "0.0~4.5, 소수 첫째 자리까지"
                        }
                      ]
                    }
                  },
                  "CONSENT_REQUIRED": {
                    "value": {
                      "code": "CONSENT_REQUIRED",
                      "message": "프로필 저장에 동의가 없음"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED · TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AUTH_REQUIRED": {
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "토큰 없음·잘못된 토큰"
                    }
                  },
                  "TOKEN_EXPIRED": {
                    "value": {
                      "code": "TOKEN_EXPIRED",
                      "message": "토큰 만료(체험 계정은 계정도 지워짐)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN_ROLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "FORBIDDEN_ROLE": {
                    "value": {
                      "code": "FORBIDDEN_ROLE",
                      "message": "역할이 맞지 않음(학생이 현황판 등)"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/explore/jobs/{jobId}/why": {
      "get": {
        "tags": [
          "탐색"
        ],
        "operationId": "getExploreWhy",
        "summary": "직무 상세 '왜 맞나요'(없으면 이때 만든다)",
        "x-auth": "STUDENT",
        "description": "권한: **STUDENT** · 화면: S4 · 규칙은 [docs/api/README.md](https://github.com/skuniv-team11/Backend/blob/develop/docs/api/README.md) '엔드포인트별 규칙'",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploreJobWhy"
                },
                "examples": {
                  "explore-why.json": {
                    "summary": "docs/api/explore-why.json",
                    "value": {
                      "jobId": 105,
                      "fit": "WEAK",
                      "why": {
                        "summary": "매장에서 손님 질문을 정리해 본 경험은 판매 데이터를 정리하는 일과 조금 이어지지만, 숫자를 다루는 일은 글에 없어 새로 배우게 돼요.",
                        "points": [
                          {
                            "text": "손님이 자주 묻는 질문을 메모해 정리한 경험은, 판매 현황을 꾸준히 기록하고 정리하는 일에 바탕이 돼요.",
                            "studentQuote": "손님이 자주 묻는 질문을 메모해 두고",
                            "jobQuote": "(가상) 판매 데이터 정리와 매출 분석 보조"
                          }
                        ],
                        "tryNew": [
                          {
                            "text": "상품별 판매 성과를 비교해 개선 방안을 내는 일을 해 볼 수 있어요.",
                            "jobQuote": "(가상) 상품별 판매 성과 분석"
                          }
                        ],
                        "prepare": {
                          "text": "엑셀로 표를 만들고 합계·평균을 내 보는 연습을 해 두면 좋아요.",
                          "jobQuote": "(가상) 판매 데이터 정리와 매출 분석 보조"
                        }
                      },
                      "fallbackReason": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED · TOKEN_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AUTH_REQUIRED": {
                    "value": {
                      "code": "AUTH_REQUIRED",
                      "message": "토큰 없음·잘못된 토큰"
                    }
                  },
                  "TOKEN_EXPIRED": {
                    "value": {
                      "code": "TOKEN_EXPIRED",
                      "message": "토큰 만료(체험 계정은 계정도 지워짐)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN_ROLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "FORBIDDEN_ROLE": {
                    "value": {
                      "code": "FORBIDDEN_ROLE",
                      "message": "역할이 맞지 않음(학생이 현황판 등)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "EXPLORE_NOT_FOUND · JOB_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "EXPLORE_NOT_FOUND": {
                    "value": {
                      "code": "EXPLORE_NOT_FOUND",
                      "message": "저장된 탐색 결과 없음"
                    }
                  },
                  "JOB_NOT_FOUND": {
                    "value": {
                      "code": "JOB_NOT_FOUND",
                      "message": "없는 직무"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "EXPLORE_NOT_CANDIDATE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "EXPLORE_NOT_CANDIDATE": {
                    "value": {
                      "code": "EXPLORE_NOT_CANDIDATE",
                      "message": "지원할 수 없거나 마감된 직무라 '왜 맞나요'를 만들지 않음"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "사람이 읽는 설명(예시 문구는 서버 문구와 다를 수 있다)"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "reason"
              ]
            },
            "description": "INVALID_INPUT일 때만"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "Explore": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "integer",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "round": {
            "$ref": "#/components/schemas/RoundRef"
          },
          "source": {
            "$ref": "#/components/schemas/ExploreSource"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExploreFallback"
              },
              {
                "type": "null"
              }
            ],
            "description": "source가 AI면 null"
          },
          "judgedWith": {
            "$ref": "#/components/schemas/ExploreJudgedWith"
          },
          "hiddenCount": {
            "type": "integer",
            "minimum": 0,
            "description": "저장한 프로필로 다시 판정해 후보에서 빠져 숨긴 직무 수"
          },
          "input": {
            "type": "object",
            "properties": {
              "experiences": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "maxItems": 3
              },
              "cards": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "maxItems": 5
              },
              "interestText": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "experiences",
              "cards",
              "interestText"
            ],
            "description": "저장한 경험 글(전화·이메일·긴 숫자는 [가림])·고른 카드 글·관심 분야"
          },
          "candidates": {
            "$ref": "#/components/schemas/ExploreCandidates"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExploreItem"
            },
            "maxItems": 5
          },
          "blockedBy": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "item": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                }
              },
              "required": [
                "item",
                "count"
              ]
            },
            "description": "items가 비었을 때만(#15와 같은 규칙)"
          }
        },
        "required": [
          "runId",
          "createdAt",
          "round",
          "source",
          "fallbackReason",
          "judgedWith",
          "hiddenCount",
          "input",
          "candidates",
          "items",
          "blockedBy"
        ]
      },
      "ExploreCandidates": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "후보 수 = eligible + needsCheck"
          },
          "eligible": {
            "type": "integer",
            "minimum": 0
          },
          "needsCheck": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "total",
          "eligible",
          "needsCheck"
        ],
        "description": "후보 = 판정이 ELIGIBLE·NEEDS_CHECK이고 기준일에 마감되지 않은 직무. INELIGIBLE·마감 직무는 AI에 보내지 않는다"
      },
      "ExploreCards": {
        "type": "object",
        "properties": {
          "round": {
            "$ref": "#/components/schemas/RoundRef"
          },
          "candidates": {
            "$ref": "#/components/schemas/ExploreCandidates"
          },
          "cards": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "pattern": "^\\d+-\\d+$",
                  "description": "카드 id `{jobId}-{순번}`"
                },
                "text": {
                  "type": "string",
                  "description": "직무 개요(또는 주차 계획) 항목 원문, 기관 이름은 '회사'로 가림"
                }
              },
              "required": [
                "id",
                "text"
              ]
            },
            "maxItems": 60,
            "description": "후보 직무를 번갈아 놓은 순서. 거의 같은 글은 하나로 합친다"
          }
        },
        "required": [
          "round",
          "candidates",
          "cards"
        ]
      },
      "ExploreFallback": {
        "type": "string",
        "enum": [
          "NO_KEY",
          "LIMITED",
          "AI_ERROR",
          "VERIFY_FAILED",
          "NO_CANDIDATES"
        ],
        "description": "`NO_KEY` AI 설정 없음 · `LIMITED` 탐색 한도 초과 · `AI_ERROR` AI 응답 없음 · `VERIFY_FAILED` AI 근거 확인 실패 · `NO_CANDIDATES` 지원할 수 있는 자리 없음"
      },
      "ExploreFit": {
        "type": "string",
        "enum": [
          "STRONG",
          "GOOD",
          "WEAK"
        ],
        "description": "`STRONG` 잘 맞음 · `GOOD` 맞음 · `WEAK` 덜 맞음"
      },
...
      "ExploreJobWhy": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "integer",
            "minimum": 1
          },
          "fit": {
            "$ref": "#/components/schemas/ExploreFit",
            "description": "탐색 결과 안이면 그 값, 밖이면 WEAK"
          },
          "why": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExploreWhyText"
              },
              {
                "type": "null"
              }
            ]
          },
          "fallbackReason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExploreFallback"
              },
              {
                "type": "null"
              }
            ],
            "description": "why가 null인 까닭(NO_KEY·LIMITED·AI_ERROR·VERIFY_FAILED)"
          }
        },
        "required": [
          "jobId",
          "fit",
          "why",
          "fallbackReason"
        ]
      },
      "ExploreJudgedWith": {
        "type": "string",
        "enum": [
          "SAVED_PROFILE",
          "RUN_PROFILE"
        ],
        "description": "`SAVED_PROFILE` 저장한 프로필로 다시 판정 · `RUN_PROFILE` 탐색 때 보낸 프로필"
      },
      "ExploreRequest": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "experiences": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 20,
              "maxLength": 200
            },
            "maxItems": 3,
            "description": "해 본 일 0~3개. 경험 1개 이상 또는 카드 3개 이상이어야 한다"
          },
          "cardIds": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^\\d+-\\d+$",
              "description": "카드 id `{jobId}-{순번}`"
            },
            "maxItems": 5,
            "description": "POST /api/explore/cards의 id, 중복 불가"
          },
          "consent": {
            "type": "boolean",
            "description": "경험 글을 AI에 보내고 결과와 함께 저장하는 데 동의. true가 아니면 400 CONSENT_REQUIRED"
          }
        },
        "required": [
          "profile",
          "experiences",
          "cardIds",
          "consent"
        ]
      },
      "ExploreSource": {
        "type": "string",
        "enum": [
          "AI",
          "RULE"
        ],
        "description": "`AI` AI 탐색 · `RULE` 규칙 추천(AI 대신)"
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "integer",
            "minimum": 1,
            "description": "`GET /api/departments`의 id"
          },
          "grade": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "completedSemesters": {
            "type": "integer",
            "minimum": 0,
            "maximum": 8
          },
          "gpa": {
            "type": "number",
            "minimum": 0,
            "maximum": 4.5,
            "description": "소수 첫째 자리까지"
          },
          "graduationExpected": {
            "type": "boolean",
            "description": "다음 졸업(2026-2 회차 → 2027년 2월) 예정이면 true"
          },
          "interestText": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "description": "적합도 추천의 질의. 200자 이하"
          },
          "homeAreaCode": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^(11|28|41)\\d{3}$",
            "description": "사는 곳(선택). null이면 통근을 서경대에서 출발로 본다. 통근 조회에만 쓴다"
          },
          "certificates": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "pattern": "^[A-Z][A-Z0-9_]*$",
              "maxLength": 40,
              "description": "`GET /api/certificates`의 code"
            },
            "maxItems": 20,
            "description": "가진 자격증(선택, ADR-0021). []이면 없음, null이거나 빠지면 답하지 않음(자격증 줄이 '직접 확인'). 판정의 자격증 줄에만 쓴다"
          }
        },
        "required": [
          "departmentId",
          "grade",
          "completedSemesters",
          "gpa",
          "graduationExpected"
        ],
        "description": "학생 프로필. 요청 본문으로만 보낸다(URL·쿼리에 넣지 않는다)"
      },
      "ProfileBody": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "profile"
        ]
      },
...
      "RoundRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "termCode": {
            "type": "string",
            "pattern": "^\\d{4}-[12]$"
          }
        },
        "required": [
          "id",
          "termCode"
        ]
      },
...
    }
  }
}